% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_xattr}
\alias{get_xattr}
\title{Retrieve the contents of the named xattr}
\usage{
get_xattr(path, name, follow_symlinks = TRUE)
}
\arguments{
\item{path}{target path (file or dir); this is auto-expanded}

\item{name}{xattr name to retrieve}

\item{follow_symlinks}{if \code{FALSE} get xattr of the symlink vs the target it references}
}
\description{
Retrieve the contents of the named xattr
}
\note{
You should really use \code{\link[=get_xattr_raw]{get_xattr_raw()}}
}
