% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{get_xattr_df}
\alias{get_xattr_df}
\title{Retrieve a data frame of xattr names, sizes and (raw) contents for a target path}
\usage{
get_xattr_df(path, follow_symlinks = TRUE)
}
\arguments{
\item{path}{path target path (file or dir); this is auto-expanded}

\item{follow_symlinks}{if \code{FALSE} get xattr of the symlink vs the target it references}
}
\description{
Retrieve a data frame of xattr names, sizes and (raw) contents for a target path
}
