% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{handle_user_prefix_param}
\alias{handle_user_prefix_param}
\title{handle_user_prefix_param}
\usage{
handle_user_prefix_param(stringvector)
}
\arguments{
\item{stringvector}{vector of strings with attribute names}
}
\value{
stringvector adjusted with required namespace (Linux)
}
\description{
handle_user_prefix_param
}
\keyword{internal}
