% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{handle_user_prefix_return}
\alias{handle_user_prefix_return}
\title{handle_user_prefix_return}
\usage{
handle_user_prefix_return(stringvector)
}
\arguments{
\item{stringvector}{vector of strings with attribute names}
}
\value{
stringvector adjusted without namespace (Linux)
}
\description{
handle_user_prefix_return
}
\keyword{internal}
