% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{has_xattrs}
\alias{has_xattrs}
\title{Test if a target path has xattrs}
\usage{
has_xattrs(path, follow_symlinks = TRUE)
}
\arguments{
\item{path}{target path (file or dir); this is auto-expanded}

\item{follow_symlinks}{if \code{FALSE} get xattr of the symlink vs the target it references}
}
\description{
Test if a target path has xattrs
}
