% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-bplist.R
\name{is_bplist}
\alias{is_bplist}
\title{Tests whether a raw vector is really a binary plist}
\usage{
is_bplist(rawv)
}
\arguments{
\item{rawv}{raw vector to test}
}
\description{
Tests whether a raw vector is really a binary plist
}
