% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{list_xattrs}
\alias{list_xattrs}
\title{List extended attribute names of a target path}
\usage{
list_xattrs(path, follow_symlinks = TRUE)
}
\arguments{
\item{path}{target path (file or dir); this is auto-expanded}

\item{follow_symlinks}{if \code{FALSE} get xattr of the symlink vs the target it references}
}
\description{
List extended attribute names of a target path
}
