% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-bplist.R
\name{read_bplist}
\alias{read_bplist}
\title{Convert binary plist to something usable in R}
\usage{
read_bplist(raw_bplist)
}
\arguments{
\item{raw_bplist}{raw vector containing the contents of a binary plist}
}
\value{
\code{list}
}
\description{
Convert binary plist to something usable in R
}
