% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xattrs-package.R
\docType{package}
\name{xattrs}
\alias{xattrs}
\alias{xattrs-package}
\title{Work With Filesystem Object Extended Attributes}
\description{
Filesystem path target extended attributes store extra, customizable, small bits of
info. For example, author name, file character encoding, short comments, security
status, etc. Methods are provided to list, extract and work with these attributes.
}
\author{
Bob Rudis (bob@rud.is)
}
