% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{get_xattr_df}
\alias{get_xattr_df}
\alias{read_xattrs}
\title{Retrieve a data frame of xattr names, sizes and (raw) contents for a target path}
\usage{
get_xattr_df(path, follow_symlinks = TRUE)

read_xattrs(path, follow_symlinks = TRUE)
}
\arguments{
\item{path}{path target path (file or dir); this is auto-expanded}

\item{follow_symlinks}{if \code{FALSE} get xattr of the symlink vs the target it references}
}
\description{
Retrieve a data frame of xattr names, sizes and (raw) contents for a target path
}
\examples{
# Create a temp file for the example
tf <- tempfile(fileext = ".csv")
write.csv(mtcars, tf)

# has attributes? (shld be FALSE)
has_xattrs(tf)
get_xattr(tf, "is.rud.setting")

# set an attribute
set_xattr(tf, "is.rud.setting.a", "first attribut")
get_xattr(tf, "is.rud.setting.a")
get_xattr_size(tf, "is.rud.setting.a")

# shld be TRUE
has_xattrs(tf)

set_xattr(tf, "is.rud.setting.b", "second attribute")
get_xattr(tf, "is.rud.setting.b")
get_xattr_size(tf, "is.rud.setting.b")

# overwrite an attribute
set_xattr(tf, "is.rud.setting.a", "first attribute")
get_xattr(tf, "is.rud.setting.a")
get_xattr_size(tf, "is.rud.setting.a")

# see all the attributes
list_xattrs(tf)

# data frame vs individual functions
get_xattr_df(tf)

# remove attribute
rm_xattr(tf, "is.rud.setting")
get_xattr(tf, "is.rud.setting")

# cleanup
unlink(tf)

}
